+++ test-utils/src/host.rs
@@ -3,14 +3,14 @@ use alloc::sync::Arc;
 use processor::{
     AdviceProvider, AdviceSource, DefaultHost, ErrorContext, MastForest, ProcessState,
 };
-use prover::{ExecutionError, Host, MemAdviceProvider, Word};
+use prover::{ExecutionError, Host, Word};
 use stdlib::{EVENT_FALCON_SIG_TO_STACK, falcon_sign};
 use vm_core::mast::MastNodeExt;
 
-pub struct TestHost(DefaultHost<MemAdviceProvider>);
+pub struct TestHost(DefaultHost);
 
 impl TestHost {
-    pub fn new(advice_provider: MemAdviceProvider) -> Self {
+    pub fn new(advice_provider: AdviceProvider) -> Self {
         Self(DefaultHost::new(advice_provider))
     }
 
@@ -18,27 +18,21 @@ impl TestHost {
         self.0.load_mast_forest(mast_forest)
     }
 
-    pub fn advice_provider(&self) -> &MemAdviceProvider {
+    pub fn advice_provider(&self) -> &AdviceProvider {
         self.0.advice_provider()
     }
 
-    pub fn advice_provider_mut(&mut self) -> &mut MemAdviceProvider {
+    pub fn advice_provider_mut(&mut self) -> &mut AdviceProvider {
         self.0.advice_provider_mut()
-    }
-
-    pub fn into_inner(self) -> MemAdviceProvider {
-        self.0.into_inner()
     }
 }
 
 impl Host for TestHost {
-    type AdviceProvider = MemAdviceProvider;
-
-    fn advice_provider(&self) -> &Self::AdviceProvider {
+    fn advice_provider(&self) -> &AdviceProvider {
         self.0.advice_provider()
     }
 
-    fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider {
+    fn advice_provider_mut(&mut self) -> &mut AdviceProvider {
         self.0.advice_provider_mut()
     }
 
@@ -79,7 +73,7 @@ impl Host for TestHost {
 ///
 /// The advice provider is expected to contain the private key associated to the public key PK.
 pub fn push_falcon_signature(
-    advice_provider: &mut impl AdviceProvider,
+    advice_provider: &mut AdviceProvider,
     process: ProcessState,
     err_ctx: &ErrorContext<'_, impl MastNodeExt>,
 ) -> Result<(), ExecutionError> {